Move client parsing of ServerHello extensions into new framework
