+++ ssl/ssl_err.c
@@ -319,6 +319,11 @@ static ERR_STRING_DATA SSL_str_functs[] = {
     {ERR_FUNC(SSL_F_TLS_PARSE_CLIENT_RENEGOTIATE),
      "tls_parse_client_renegotiate"},
     {ERR_FUNC(SSL_F_TLS_PARSE_CLIENT_USE_SRTP), "tls_parse_client_use_srtp"},
+    {ERR_FUNC(SSL_F_TLS_PARSE_SERVER_KEY_SHARE),
+     "tls_parse_server_key_share"},
+    {ERR_FUNC(SSL_F_TLS_PARSE_SERVER_RENEGOTIATE),
+     "tls_parse_server_renegotiate"},
+    {ERR_FUNC(SSL_F_TLS_PARSE_SERVER_USE_SRTP), "tls_parse_server_use_srtp"},
     {ERR_FUNC(SSL_F_TLS_POST_PROCESS_CLIENT_HELLO),
      "tls_post_process_client_hello"},
     {ERR_FUNC(SSL_F_TLS_POST_PROCESS_CLIENT_KEY_EXCHANGE),
